[file-system][Android] Add `contentUri` property.
