+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/unifiedfile/UnifiedFileInterface.kt
@@ -1,6 +1,7 @@
 package expo.modules.filesystem.unifiedfile
 
 import android.net.Uri
+import expo.modules.kotlin.AppContext
 
 interface UnifiedFileInterface {
   fun exists(): Boolean
@@ -17,6 +18,7 @@ interface UnifiedFileInterface {
   fun lastModified(): Long?
   val creationTime: Long?
   val fileName: String?
+  fun getContentUri(appContext: AppContext): Uri
   fun outputStream(): java.io.OutputStream
   fun inputStream(): java.io.InputStream
   fun length(): Long
